rename folder component
